refactor(cli): extract the pure run-row model out of the Ink components
The web console has to draw the same rows as the terminal — same glyph column,
same block bar, same trailing number, same run outcome. Those rules lived
inside Ink components, so a browser renderer would have had to restate them and
they would have drifted the first time either side changed.

`src/ui/model/` now owns them and imports no renderer; a test in that directory
fails if anything there imports ink or react. Every previous home re-exports,
so no other import moved.
