+++ examples/cli/src/ui/components/TaskDetailColumn.tsx
@@ -7,28 +7,12 @@
 import { Box, Text } from "ink";
 import React from "react";
 import { useCliTheme } from "../CliThemeContext";
-import { formatCliDuration } from "../formatCliDuration";
+import { taskDetailText } from "../model/runRowModel";
 
 /** Column width holds the widest value either state produces (`847ms`, `100%`). */
 export const TASK_DETAIL_COLUMN_WIDTH = 6;
 
-/**
- * The one number at the end of a task row: how far along it is while it runs,
- * and how long it took once it settles. Two states of one column rather than
- * two columns, because only one of them is ever true, and a fixed width keeps
- * the values aligned down the run.
- */
-export function taskDetailText(
-  progress: number | undefined,
-  durationMs: number | undefined,
-  running: boolean
-): string {
-  if (running) {
-    if (progress === undefined) return "";
-    return `${Math.round(Math.max(0, Math.min(100, progress)))}%`;
-  }
-  return durationMs === undefined ? "" : formatCliDuration(durationMs);
-}
+export { taskDetailText };
 
 export function TaskDetailColumn({
   progress,
